refactor(rollup): some changes
